Se logran establecer consumir servicios post(crear) y get(listar) y se establecen dos de las reglas de negocio(pago máximo permitido por cancha y abono mínimo permitido del 50% según cancha)

+++ java-arquitectura-hexagonal/microservicio/dominio/src/main/java/com/ceiba/reserva/puerto/repositorio/RepositorioReserva.java
@@ -9,14 +9,14 @@ public interface RepositorioReserva {
      * @param reserva
      * @return el id generado
      */
-    int crear(Reserva reserva);
+    Long  crear(Reserva reserva);
 
     /**
      * Verifica si existe reserva asociada al idReserva
      * @param idReserva
      * @return
      */
-    boolean existe(int idReserva);
+    boolean existe(Long idReserva);
 
     /**
      * Permite actualizar los datos de un reserva
@@ -29,6 +29,6 @@ public interface RepositorioReserva {
      * Permite eliminar el registro de una reserva por medio de su id
      * @param idReserva
      */
-    void eliminar(int idReserva);
+    void eliminar(Long idReserva);
 
 }
