+++ java-arquitectura-hexagonal/microservicio/dominio/src/main/java/com/ceiba/reserva/servicio/ServicioCrearReserva.java
@@ -16,7 +16,7 @@ public class ServicioCrearReserva {
         this.repositorioReserva = repositorioReserva;
     }
 
-    public int ejecutar(Reserva reserva) {
+    public Long ejecutar(Reserva reserva) {
         validarExistenciaPrevia(reserva);
         return this.repositorioReserva.crear(reserva);
     }
